### 2017-07-12 #316 @arfon
> @whedon start review

@GabrieleRovigatti - we'll need one of the JOSS editors to volunteer to edit this submission before we can proceed.
