### 2017-08-02 #316 @arfon
@GabrieleRovigatti - yes, I'm sorry this is taking a while. We don't really have anyone with economics background on the editorial board currently. If you're willing to wait I'm happy to go and try and search out someone.
