### 2017-08-02 #316 @GabrieleRovigatti
@arfon I am willing to wait as long as it is necessary! I am sorry if this is going to take long / will need a supplementary load of work for you. I like your journal, its 'mission' and main goals, but maybe I should have read more carefully the editorial board background before the submission. Hopefully you will find someone soon - tip: most people working in the relevant literature use Stata for panel-data analyses like the one I propose, maybe you may try among Stata-users?

By the way, sorry for mispelling your name (twice!) previously, I just figured out the exact spelling with your last email.
