### 2017-08-10 #316 @arfon
> @arfon I am willing to wait as long as it is necessary! I am sorry if this is going to take long / will need a supplementary load of work for you. I like your journal, its 'mission' and main goals, but maybe I should have read more carefully the editorial board background before the submission.

Thanks for your patience. We're shortly going to be adding a new editor to help with submissions like yours :-)

> By the way, sorry for mispelling your name (twice!) previously, I just figured out the exact spelling with your last email.

No problem. It's a common mistake!
