### 2017-08-12 #316 @arfon
@leeper - would you be able to edit this submission for JOSS?
