### 2017-08-13 #316 @leeper
@GabrieleRovigatti I'm not seeing a `paper.md` file in the repository. Can you add one per [the author guidelines](http://joss.theoj.org/about#author_guidelines)? An example paper is [available here](https://github.com/arfon/fidgit/tree/master/paper) if you'd like to see what we have in mind.
