### 2017-08-13 #316 @arfon
> @GabrieleRovigatti I'm not seeing a paper.md file in the repository. Can you add one per the author guidelines? An example paper is available here if you'd like to see what we have in mind.

Thanks @leeper. Also, the section [_'what should my paper contain'_](http://joss.theoj.org/about#paper_structure) here has some guidance for the sort of content we're looking to be present in the `paper.md`.

Many thanks!
