+++ prodest/paper/prodest.bib
@@ -1,0 +1,52 @@
+
+@Article{ackerberg_etal15,
+  Title                    = {Identification Properties of Recent Production Function Estimators},
+  Author                   = {Ackerberg, D. and Caves, K. and Frazer, G.},
+  Journal                  = {Econometrica},
+  Year                     = {2015},
+  Number                   = {6},
+  Pages                    = {2411-2451},
+  Volume                   = {83}, 
+  doi 						= {10.3982/ECTA13408}
+}
+
+@Article{levinsohn_etal03,
+  Title                    = {Estimating Production Functions Using Inputs to Control for Unobservables},
+  Author                   = {Levinsohn, J. and Petrin, A.},
+  Journal                  = {The Review of Economic Studies},
+  Year                     = {2003},
+  Number                   = {2},
+  Pages                    = {317-341},
+  Volume                   = {70}, 
+  doi 						= {10.1111/1467-937X.00246}
+}
+
+@Article{olley_etal96,
+  Title                    = {The Dynamics of Productivity in the Telecommunications Equipment Industry},
+  Author                   = {Olley, S. G. and Pakes, A.},
+  Journal                  = {Econometrica},
+  Year                     = {1996},
+  Number                   = {6},
+  Pages                    = {1263-1297},
+  Volume                   = {64},
+  doi 						= {10.2307/2171831}
+}
+
+@Article{wooldridge_09,
+  author={Wooldridge, J. M.},
+  title={On estimating firm-level production functions using proxy variables to control for unobservables},
+  journal={Economics Letters},
+  year={2009},
+  volume={104},
+  pages={112-114},
+  doi = {10.1016/j.econlet.2009.04.026}
+}
+
+@article{rovigatti_17,
+  title={Production Function Estimation in R: The prodest package.},
+  author={Rovigatti, G},
+  journal={R package},
+  Version = {0.1.1}
+  url = {https://cran.r-project.org/package=prodest.}
+}
+
