+++ prodest/paper/prodest.bib
@@ -50,3 +50,10 @@
   url = {https://cran.r-project.org/package=prodest.}
 }
 
+@online{prodestGit_17,
+  title={prodest},
+  author={Rovigatti, G},
+  url = {https://github.com/GabrieleRovigatti/prodest/tree/master/prodest}
+}
+
+
