### 2017-08-16 #316 @GabrieleRovigatti
@leeper sorry for that, I was sure to have already uploaded the paper and the .bib file. Hopefully, now you can find everything [here](https://github.com/GabrieleRovigatti/prodest/tree/master/prodest/paper)

@arfon I tried to follow the guidelines and - at the same time - to be as concise as possible, hoping that interested readers will refer to the vignette included in the package itself.
