### 2017-08-16 #316 @leeper
Thanks for your submission @GabrieleRovigatti. I've confirmed that this submission has:

- [X]  An OSI approved license (GPL)
- [X] A description of the software and its research applications
- [X] Is in scope for this journal

Two things:

 1. I might be inclined to add just a sentence or two of additional text to the paper.md to contextualize the software for a general audience that may not know what Total Factor Productivity (TFP) estimation is.
 2. Are you submitting both the R and Stata versions as one package or just the R version (as linked above)?

@sebastianbarfort has agreed to review.
