### 2017-08-17 #316 @GabrieleRovigatti
@leeper thanks for having started the review - and for accepting the task in the first place! And thanks @sebastianbarfort, too, for his time and effort.

1. I added a sentence explaining - in a word - what is meant by TFP.

2. I am submitting the R routine only. I already submitted the Stata version to the Stata Journal. By the way, I am not sure Stata routines comply with the 'open source' nature of the journal, or would you accept proprietary software packages, too?
