### 2017-08-17 #316 @leeper
@GabrieleRovigatti Okay, great. Let's proceed on R only.

Side note: @arfon @karthik we have a good question on journal scope. If the software is open source but it's an add-on to a proprietary tool (like Stata) is that in or out of scope for JOSS?
