### 2017-08-17 #371 @whedon
[REVIEW]: Production Function Estimation in R: The prodest Package
**Submitting author:** @GabrieleRovigatti (<a href="http://orcid.org/0000-0003-4021-3070">Gabriele Rovigatti</a>)
**Repository:** <a href="https://github.com/GabrieleRovigatti/prodest" target ="_blank">https://github.com/GabrieleRovigatti/prodest</a>
**Version:** v0.1.1
