### 2017-08-17 #316 @leeper
@sebastianbarfort Review checklist is up at: https://github.com/openjournals/joss-reviews/issues/371
