### 2017-08-17 #316 @karthik
@leeper

We did have a very long thread about this (I'll forward it to you). But the outcome was:

> We strongly prefer software that doesn't rely upon proprietary environments.
Open source submissions that do rely upon a proprietary environment/tool are acceptable provided that it is possible to find a reviewer for the submission and that the submitting author is responsible for finding these reviewers.
Add to the reviewer guidelines:
That the reviewer should encourage the author to consider porting their tools to be compatible with open source/free variants of the runtime e.g. Octave instead of MATLAB. (This won't be a condition of acceptance however).
