### 2017-08-17 #316 @leeper
@karthik Thanks. I'll close this then since the review issue is now open and all seems well here.
