+++ prodest/README.md
@@ -17,4 +17,22 @@ Rovigatti, G. (2017).
 Production Function Estimation in R: The prodest Package.  
 Working paper.      
  
-To install the latest stable version of `prodest` in R type `install.packages("prodest")'.
+Installation: To install the latest stable version of `prodest` in R type `install.packages("prodest")`.
+
+Example: In R, type 
+
+` require(prodest)
+  data(chilean) ## Chilean data on production.
+  #we fit a model with two free (skilled and unskilled), one state (capital) and one proxy variable (electricity)
+  #with two different optimizers
+  LP.fit <- prodestLP(chilean$Y, fX = cbind(chilean$fX1, chilean$fX2), chilean$sX,
+                        chilean$pX, chilean$idvar, chilean$timevar, seed = 154673)
+  LP.fit.solnp <- prodestLP(chilean$Y, fX = cbind(chilean$fX1, chilean$fX2), chilean$sX,
+                        chilean$pX, chilean$idvar, chilean$timevar, opt = 'solnp')
+  #show results
+  summary(LP.fit)
+  summary(LP.fit.solnp)
+
+  #show results in .tex tabular format
+  printProd(list(LP.fit, LP.fit.solnp))
+ `
