### 2017-09-21 11c57a1
Create contr_guidelines.md
