+++ contr_guidelines.md
@@ -1,0 +1,38 @@
+# How to contribute to prodest
+
+There are several ways to contribute to an open source project. If you are reading this, you are probably interested in 
+contributing to prodest: how nice of you! In order to ensure a) not to waste your time, b) maximize the efficacy of your 
+contribution, and c) set some contribution standards, below is a list of best practices for prodest contributors.
+
+#### **Did you find a bug?** ####
+
+* **Ensure the bug was not already reported** by looking at the prodest [Issues](https://github.com/GabrieleRovigatti/prodest/issues) page.
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/GabrieleRovigatti/prodest/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** reproducing the issue. In particular, if that is found in the **Stata** code, please add a logfile featuring a **creturn list** output in order to enhance reproducibility. The first best would be to reproduce the bug with the data provided with the packages: that is to say the Chilean data downloadable from prodest helpfile (Stata) or the *data(chilean)* included in the R package.
+
+#### **Did you write a patch that fixes a bug?** ####
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure to feature the pull request with a clear description of the problem and the proposed solution. Include the relevant issue number when applicable.
+
+#### **Did you fix the code according to some standard coding conventions - i.e., cosmetic patches?** ####
+
+* All cosmetic changes will be considered, but guys: would you really waste your time going through a code line by line and opening a pull request without adding anything substantial to the project? 
+
+#### **Do you intend to develop a new feature / add a new function / change an existing one?** ####
+
+* Suggest your changes to gabriele.rovigatti@gmail.com - in order to avoid duplicates, maybe we are already working on it! - then start writing your own function.
+
+* Do not open an issue on GitHub until you have collected positive feedback about the change. GitHub issues are primarily intended for bug reports and fixes.
+
+#### **Do you have questions about the source code?** ####
+
+* For any questions regarding packages and functions read the helpfile (Stata version) or the package documentation (R). In case of very hard-to-find answers, send your inquiries to gabriele.rovigatti@gmail.com
+
+Thanks!
+
+Gabriele Rovigatti
+
+
+
