+++ README.md
@@ -1,0 +1,3 @@
+# prodest
+
+Stata and R packages for Production Function Estimation. 
