+++ README.md
@@ -1,3 +1,19 @@
 # prodest
 
 Stata and R packages for Production Function Estimation. 
+
+`prodest` ([Rovigatti, 2017](https://CRAN.R-project.org/package=prodest)) implements most of the methods for production function estimation. The prodest package provides functions to simulate panel data, estimate the productivity, state and free variables parameters. Full description of the models, their issues and characteristics is available in [Mollisi et al., 2017](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2916753).
+
+The latest stable version of `prodest` is available on CRAN at 'https://cran.r-project.org/package=prodest'.
+
+The latest development version of `prodest` is available at 'https://github.com/GabrieleRovigatti/prodest/tree/master/prodest'.
+
+Please cite `prodest` in publications:
+
+Rovigatti, G. (2017).  
+_prodest: Production Function Estimation in R_.  
+R package.  
+
+Rovigatti, G. (2017).  
+Production Function Estimation in R: The prodest Package.  
+Working paper.      
