+++ README.md
@@ -2,13 +2,17 @@
 
 Stata and R packages for Production Function Estimation. 
 
-`prodest` ([Rovigatti, 2017](https://CRAN.R-project.org/package=prodest)) implements most of the methods for production function estimation. The prodest package provides functions to simulate panel data, estimate the productivity, state and free variables parameters. Full description of the models, their issues and characteristics is available in [Mollisi et al., 2017](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2916753).
+`prodest` in Stata ([Mollisi et al., 2017],(https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2916753)) is available on the SSC archive - type `ssc install prodest` to install the latest stable version.
 
-The latest stable version of `prodest` is available on CRAN at 'https://cran.r-project.org/package=prodest'.
+`prodest` in R ([Rovigatti, 2017](https://CRAN.R-project.org/package=prodest)) is available on the CRAN -  type `install.packages("prodest")` to install the latest stable version. 
 
-The latest development version of `prodest` is available at 'https://github.com/GabrieleRovigatti/prodest/tree/master/prodest'.
+The packages implement most of the methods for production function estimation. The R prodest package provides functions to estimate the productivity, state and free variables parameters. Full description of the models, their issues and characteristics is available on both the packages' documentation and the relative working papers.
 
 Please cite `prodest` in publications:
+
+Mollisi, V. and Rovigatti, G. (2017).
+_Theory and Practice of TFP Estimation: The Control Function Approach Using Stata_.
+CEIS working paper no. 399
 
 Rovigatti, G. (2017).  
 _prodest: Production Function Estimation in R_.  
