### 2017-09-26 #371 @leeper
@sebastianbarfort are you satisfied with the responses to issues you raised? Or are further changes needed?
