### 2017-09-27 #371 @sebastianbarfort
I'm super satisfied. Thanks for quick responses @GabrieleRovigatti, it's a great package!
