### 2017-10-07 #371 @arfon
@GabrieleRovigatti - I can't find the paper (`paper.md`) associated with this submission in the repository. Can you point me to it please?
