### 2017-10-07 #371 @arfon
@GabrieleRovigatti if [this](https://github.com/GabrieleRovigatti/prodest/blob/master/prodest/paper/prodest.md) is the paper, please rename these files to `paper.md` and `paper.bib`
