### 2017-10-09 #371 @arfon
@GabrieleRovigatti - please merge this PR https://github.com/GabrieleRovigatti/prodest/pull/6 and then could you please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
