### 2017-10-09 #371 @GabrieleRovigatti
@arfon the DOI is  10.5281/zenodo.1004765  - how do I update the thread with it? thanks
