### 2017-10-09 #371 @GabrieleRovigatti
@arfon, @leeper, @sebastianbarfort thank you all guys for the review process, it has been fruitful for the command implementation and interesting for me to see how you work.
